changed some stuff regarding the classes, changed the relationship between userAttack and Round

+++ src/main/java/com/example/demo/user/User.java
@@ -20,8 +20,8 @@ public class User {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long userId;
 
-    private int score;
-    private int health = 100;
+    private Long score;
+    private Long health = 100l;
 
     private String username;
     private String password;
